Update Folder Structure and Build Index Reexports
