+++ src/components/global/index.ts
@@ -1,0 +1,6 @@
+// Reexport Modals
+export { UserAuth } from './modal/auth/UserAuth';
+export { AddEmployee } from './modal/employee/AddEmployee';
+
+// Reexport Wrapper
+export { GlobalWrapper } from './wrapper/GlobalWrapper';
