+++ src/components/global/wrapper/type/AuthWrapper.tsx
@@ -1,8 +1,7 @@
 import React, { useEffect } from 'react';
-
 import { useDispatch } from 'react-redux';
-import { userAccount } from '../../redux';
-import { auth } from '../../config/firebase';
+import { userAccount, modalOpen, modalClose } from '../../../../redux';
+import { auth } from '../../../../config/firebase';
 
 // Modal Wrapper
 const AuthWrapper: React.FC = ({ children }) => {
@@ -11,8 +10,11 @@ const AuthWrapper: React.FC = ({ children }) => {
   const accountChange = () => {
     const unsub = auth.onAuthStateChanged((user) => {
       if (user) {
+        dispatch(modalClose({ modalName: 'login' }));
         dispatch(userAccount({ account: user }));
         console.log(`Account changed to ${user.email}`);
+      } else {
+        dispatch(modalOpen({ modalName: 'login' }));
       }
     });
