+++ src/components/global/wrapper/type/ModalWrapper.tsx
@@ -1,15 +1,14 @@
 import React from 'react';
 
 // Import Modals
-import { AddEmployee, LoginUser, RegisterUser } from '../modals';
+import { AddEmployee, UserAuth } from '../../../global';
 
 // Modal Wrapper
 const ModalWrapper: React.FC = ({ children }) => {
   return (
     <>
       {children}
-      <LoginUser />
-      <RegisterUser />
+      <UserAuth />
       <AddEmployee />
     </>
   );
