--- src/components/modals/index.ts
@@ -1,3 +1,0 @@
-export { LoginUser } from './option/LoginUser';
-export { RegisterUser } from './option/RegisterUser';
-export { AddEmployee } from './employee/AddEmployee';
