+++ src/components/pages/employee/EmployeeFooter.tsx
@@ -3,7 +3,7 @@ import { IonButton, IonCol, IonGrid, IonLabel, IonRow } from '@ionic/react';
 import { useDispatch } from 'react-redux';
 import { modalOpen, emplFetch } from '../../../redux';
 
-const EmployeeFooter: React.FC = () => {
+export const EmployeeFooter: React.FC = () => {
   const dispatch = useDispatch();
 
   return (
@@ -35,5 +35,3 @@ const EmployeeFooter: React.FC = () => {
     </IonGrid>
   );
 };
-
-export default EmployeeFooter;
