--- src/components/pages/option/AccountSection.css
@@ -1,5 +1,0 @@
-.background {
-  background-color: var(--ion-toolbar-background);
-  color: var(--ion-text-color);
-  margin: 15px 0px;
-}
