+++ src/components/pages/option/auth/AccountSection.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import { IonCard } from '@ionic/react';
+import AccountCard from './AccountCard';
+
+export const AccountSection: React.FC = () => {
+  return (
+    <IonCard>
+      <AccountCard />
+    </IonCard>
+  );
+};
