+++ src/components/shared/Page/PageContainer.tsx
@@ -1,11 +1,24 @@
 import React from 'react';
-import { IonPage } from '@ionic/react';
-import { PageContainerProps } from './pageTypes';
+import {
+  IonButton,
+  IonButtons,
+  IonContent,
+  IonFooter,
+  IonGrid,
+  IonHeader,
+  IonPage,
+  IonTitle,
+  IonToolbar,
+} from '@ionic/react';
+import { useDispatch } from 'react-redux';
+import { modalClose } from '../../../redux';
+import {
+  PageContainerProps,
+  PageContentProps,
+  PageHeaderProps,
+} from './pageTypes';
 
-// Import Components
-import PageContent from './PageContent';
-
-const PageContainer: React.FC<PageContainerProps> = ({
+export const PageContainer: React.FC<PageContainerProps> = ({
   pageTitle,
   color,
   fullscreen,
@@ -14,7 +27,13 @@ const PageContainer: React.FC<PageContainerProps> = ({
   children,
 }) => {
   const content = (
-    <PageContent pageTitle={pageTitle} color={color} fullscreen={fullscreen} modal={modal} footer={footer}>
+    <PageContent
+      pageTitle={pageTitle}
+      color={color}
+      fullscreen={fullscreen}
+      modal={modal}
+      footer={footer}
+    >
       {children}
     </PageContent>
   );
@@ -22,4 +41,57 @@ const PageContainer: React.FC<PageContainerProps> = ({
   return <>{modal ? content : <IonPage>{content}</IonPage>}</>;
 };
 
-export default PageContainer;
+const PageContent: React.FC<PageContentProps> = ({
+  pageTitle,
+  color,
+  fullscreen,
+  modal,
+  footer,
+  children,
+}) => (
+  <>
+    <PageHeader pageTitle={pageTitle} color={color} modal={modal} />
+    <IonContent fullscreen={fullscreen}>
+      <PageHeader pageTitle={pageTitle} color={color} condense />
+      <IonGrid className="ion-no-padding" fixed>
+        {children}
+      </IonGrid>
+    </IonContent>
+    {footer && <IonFooter className="ion-no-border">{footer}</IonFooter>}
+  </>
+);
+
+const PageHeader: React.FC<PageHeaderProps> = ({
+  pageTitle,
+  color,
+  condense,
+  modal,
+}) => {
+  const dispatch = useDispatch();
+
+  const handleClose = () => {
+    if (modal) {
+      dispatch(modalClose({ modalName: modal }));
+    }
+  };
+
+  return (
+    <IonHeader
+      collapse={condense ? 'condense' : undefined}
+      translucent={!condense}
+    >
+      <IonToolbar>
+        <IonTitle color={color} size={condense ? 'large' : undefined}>
+          {pageTitle}
+        </IonTitle>
+        {modal && !condense && (
+          <IonButtons slot="secondary">
+            <IonButton color="danger" onClick={handleClose}>
+              Close
+            </IonButton>
+          </IonButtons>
+        )}
+      </IonToolbar>
+    </IonHeader>
+  );
+};
