+++ src/components/shared/index.ts
@@ -1,0 +1,2 @@
+export { EmptyPage } from './empty/EmptyPage';
+export { PageContainer } from './page/PageContainer';
