+++ src/pages/SchedulePage.tsx
@@ -5,15 +5,19 @@ import ScheduleProvider from '../hooks/ScheduleContext';
 import WeekProvider from '../hooks/WeekContext';
 
 // Import Components
-import PageContainer from '../components/shared/Page/PageContainer';
-import ViewDisplay from '../components/pages/schedule/ViewDisplay';
-import ScheduleFooter from '../components/pages/schedule/ScheduleFooter';
+import { ScheduleFooter, ViewDisplay } from '../components/pages';
+import { PageContainer } from '../components/shared';
 
 const SchedulePage: React.FC = () => {
   return (
     <ScheduleProvider>
       <WeekProvider>
-        <PageContainer pageTitle="Schedule" color="schedule" footer={<ScheduleFooter />} fullscreen>
+        <PageContainer
+          pageTitle="Schedule"
+          color="schedule"
+          footer={<ScheduleFooter />}
+          fullscreen
+        >
           <ViewDisplay />
         </PageContainer>
       </WeekProvider>
