+++ src/pages/TaskPage.tsx
@@ -1,13 +1,11 @@
 import React from 'react';
 
 // Import Components
-import PageContainer from '../components/shared/Page/PageContainer';
+import { PageContainer } from '../components/shared';
 
 const TaskPage: React.FC = () => {
   return (
-    <PageContainer pageTitle="Task" color="task" fullscreen>
-
-    </PageContainer>
+    <PageContainer pageTitle="Task" color="task" fullscreen></PageContainer>
   );
 };
