+++ src/redux/index.ts
@@ -1,19 +1,29 @@
+// Root Reexport
 export type { RootState } from './rootReducer';
 
-export {
-  userAccount,
-  userInput,
-  userLogin,
-  userRegister,
-  userLogout,
-} from './authentication/authActions';
+// Authentication Reexports
+export { userAccount, userInput } from './authentication/authActions';
 
+export {
+  userLogin,
+  userLogout,
+  userRegister,
+} from './authentication/thunks/userThunks';
+
+export {
+  accountUpdate,
+  accountDelete,
+} from './authentication/thunks/accountThunks';
+
+// Employee Reexports
+export { emplSet, emplFetch, emplAdd } from './employee/employeeAction';
+
+// Modal Reexports
 export { modalOpen, modalClose } from './modal/modalActions';
 
 export type { Modals } from './modal/modalTypes';
 
+// Notification Reexports
 export { notifSend, notifDismiss } from './notification/notifActions';
 
 export type { NotifType } from './notification/notifTypes';
-
-export { emplSet, emplFetch, emplAdd } from './employee/employeeAction';
